+++ chromium/net/reporting/reporting_test_util.cc
@@ -334,12 +334,12 @@ void TestReportingService::ProcessHeader(const GURL& url,
 }
 
 void TestReportingService::RemoveBrowsingData(
-    int data_type_mask,
+    uint64_t data_type_mask,
     const base::RepeatingCallback<bool(const GURL&)>& origin_filter) {
   NOTREACHED();
 }
 
-void TestReportingService::RemoveAllBrowsingData(int data_type_mask) {
+void TestReportingService::RemoveAllBrowsingData(uint64_t data_type_mask) {
   NOTREACHED();
 }
